cramfs: support 36-bit physical sapce
Current fsl_p4080 run with 36-bit physical address so we should extend
those 32-bit variables to identify correctly actual address. As so we
can access cramfs with map/read as normal.

Remove one reduplicate kmap and pass the appropriate virtual address to
flush_tlb_page() when map cramfs .

+++ include/linux/cramfs_fs_sb.h
@@ -11,7 +11,7 @@ struct cramfs_sb_info {
 			unsigned long files;
 			unsigned long flags;
 #ifdef CONFIG_CRAMFS_LINEAR
-			unsigned long linear_phys_addr;
+			unsigned long long linear_phys_addr;
 			char	     *linear_virt_addr;
 #endif /* CONFIG_CRAMFS_LINEAR */
 };
